Trial refactor complete and integrated into rest of code

+++ app/src/main/java/com/example/appraisal/backend/trial/MeasurementTrial.java
@@ -32,7 +32,7 @@ public class MeasurementTrial implements Trial {
      * @param measurement
      *      The measurement obtained
      */
-    public void setMeasurement(float measurement) {
+    public void setMeasurement(double measurement) {
         this.measurement = measurement;
     }
